Update ardreadV4.py
Added the ability for the Python program to detect an Arduino Mega on a Mac and switch the port accordigly.

Also fixed errors with the code freezing on a read because the arduino missed a command.

Prints writing on the terminal when a keyboard interrupt is detected to tell the user that the excel file is being written.