Update SensorSHT35Reading.ino
Added the ability for the arduino to detect if there is no command for 5 seconds. It will then arbitrarily send data from sensor one and wait another 5 seconds. This was done so that the python code wouldn't be waiting for the arduino to output something if the arduino missed the command.